Fix FEL boot flow, Clovershell protocol, and kernel operations
Major fixes verified with real SFC Mini hardware:

- Fix AWUSBRequest struct: 16→32 bytes matching sunxi-tools wire format
- Fix FEL memory addresses: U-Boot→0x47000000, boot.img→0x47400000
- Use FES1 binary at 0x2000 for DRAM init (not SPL from uboot.bin)
- Add U-Boot bootcmd patching (boota 47400000) for RAM boot
- Rewrite ClovershellClient with real protocol (4-byte header packets,
  kill-all init, multi-packet USB transfers, session-based exec)
- Fix Clovershell VID/PID to 0x1F3A:0xEFE8 (set by gadget driver)
- Fix KernelManager to use sunxi-flash read_boot2/burn_boot2 for NAND
- Add fes1.bin resource and HakchiResources.getFES1()
- Patch boot.img cmdline for hakchi-clovershell mode on the fly
- Add new feature modules: Cheats, FTP, NAND, RetroArch, SaveState,
  Scraper, TaskQueue, FolderManager, BoxArt, GameExporter
- Bundle boot resources (boot.img, uboot.bin, fes1.bin)
- Include tested C tools (fel-boot.c, clover-exec.c)
- Clean up compiled binaries and obsolete tools from repo

+++ Sources/Hakchi/Core/Shell/SSHClient.swift
@@ -112,8 +112,8 @@ final class SSHClient {
                 session,
                 channelType,
                 UInt32(channelType.count),
-                UInt32(2 * 1024 * 1024),  // LIBSSH2_CHANNEL_WINDOW_DEFAULT
-                UInt32(32768),            // LIBSSH2_CHANNEL_PACKET_DEFAULT
+                2 * 1024 * 1024,  // LIBSSH2_CHANNEL_WINDOW_DEFAULT
+                32768,            // LIBSSH2_CHANNEL_PACKET_DEFAULT
                 nil, 0
             ) else {
                 continuation.resume(throwing: HakchiError.sshConnectionFailed("Failed to open channel"))
